Clarify source of IWebHostEnvironment
Removed IWebHostEnvironment parameter from Configure method to make it clear that the class-level property injected into the Constructor is the same instance used in Configure. 
Renamed env to HostEnvironment for convention's sake, but also, if it's going to be exposed publicly, it makes sense to have a more meaningful name.